+++ src/solutions/00001-000100/00031-00040/index.mjs
@@ -1,7 +1,9 @@
 import * as problem00031 from './00031';
 import * as problem00032 from './00032';
+import * as problem00033 from './00033';
 
 export default {
   problem00031,
   problem00032,
+  problem00033,
 };